refactor: Implement Magnetic Dock system and overhaul UI components
- Replaced `FocusLayoutHubComponent` with `MagneticDockHubComponent`, utilizing interactive `MagneticDock` panels for stats and relationships that expand on hover/click.
- Implemented dynamic layout logic to resize and center the `EventCardComponent` based on dock expansion states.
- Overhauled `ChoiceButtonComponent` with `SynColors` styling, `HoverCallbacks`, and a custom `setOpacity` method for fade effects.
- Refactored `EventCardComponent` to use `MoveEffect` and `_CustomOpacityEffect` for staggered button entrances, removing the obsolete `_TappableButtonWrapper`.
- Updated `SynGame` router to instantiate the new `MagneticDockHubComponent`.
- Removed redundant type cast in `TopBarComponent` flash logic.

+++ flutter/lib/components/ui/panels/top_bar_component.dart
@@ -273,7 +273,7 @@ class TopBarComponent extends PositionComponent with HasGameReference<SynGame> {
 
     if (_flashAlpha > 0) {
       _flashAlpha =
-          (_flashAlpha - dt * _kFlashFadePerSecond).clamp(0.0, 1.0) as double;
+          (_flashAlpha - dt * _kFlashFadePerSecond).clamp(0.0, 1.0);
     }
   }
